LEP-7660
Removed expando-attribute

Removed boolean[] and Date[] from the jsp and action (cause it was confusing, need to rework this).

Renamed a bunch of Language keys.

Made it so that when you edit custom attribute, you're only changing the default and not the actual value. Jorge will pick this up and move the button somewhere else besides being right next to the user.

git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@21011 05bdf26c-840f-0410-9ced-eb539d925f36