LPS-32223 This is very different from the previous commit. Now, we're optimizing for readability and "correct behavior". Prior to these changes, I don't think this code actually works.
The first line was: long[] categorizableGroupIds = new long[groupIds.length];

That creates an array of categorizableGroupIds, but, it's just empty with a bunch of 0s. So if the number of groupIds is 5. You have an array of {0, 0, 0, 0, 0}

Then, you check if that array has the group ID, and if it doesn't , you append. So you end up with

{0, 0, 0, 0, 0, 538}  where 538 is the next group ID. I don't think that was the intended logic.

My new changes ensures that there are no duplicate group IDs. But, it costs me a HashSet and the conversion from as HashSet to a Long[] to a long[]. I'm willing to pay this price because the ArrayUtil.contains is an iteration anyways, compared to the HashSet.contains. So it's about the same price I think...

Anyways, the other logic was wrong, so it needed to be fixed.