LPS-32362 Declare supported classNames in FolderSearcher.
While the FacetSearcher fully overrides the BaseIndexer.createFullQuery method, the FolderSearcher simply extends it. Thus, if FolderSearcher.getClassNames returns null, a NullPointerException will be thrown by the BaseIndexer.createFullQuery method. It is assumed that currently only folders of the declared classNames will be supported, since for them the classPK is unique (they use the same counter). This means this feature is not yet supported for folder entities created in plugins.