LPS-33093 Since we're heavily changing these classes, go for the more modern pattern:
String[] xyz = {'abc'};

String[] xyz =
	{'abc', 'def'};

String[] xyz = {
	'abc', 'def', 'ghi'
};

All 3 are good. I use the first one if I can fit it in line 80. I use the second one if it can fit in line 80 after one new line. And if it still won't fit, I go for the big block. We've been slowly converting to this the last few years already (so this pattern is not that new).