LPS-34520 Remove the ability to provide different help message per Application Display Template type
This is not so useful now that we have the palette and it' s just adding noise (and more maintenance work). I have decided to leave it in the framework so that third party developers still have the freedom to provide their own custom help messages for their template types.