LPS-45170 Move the logic to the LayoutService.
Notice that the original logic in SitesImpl was not correct since it only considered the case where layoutSet.getPageCount() == 1.
However it may happen that the getPageCount() > 1 with only one root layout (e.g. the root layout has children). In that case, the original logic would allow removing the root layout and cause LPS-26732 again.
Now the logic checks if it's the only public root layout of the default site.