LPS-45171 Display unambiguous vocabulary title in categories navigation and categories summary.
The site descriptive name is displayed for inherited vocabularies when there is a vocabulary in the current site with the same title.
This does not apply to the vocabulary selector (for administrators), where the site name is always shown for inherited vocabularies, regardless if there is a vocabulary with the same title in the current site.
This allows administrators to know in which site they can edit that vocabulary.