LPS-49277 Split DefaultDLFileVersionDisplayContextHelper in two: FileEntryDisplayContextHelper and FileVersionDisplayContextHelper. Then, spread methods from the previous file to the new ones depending on what they operate (file or version).
The motivation for this helpers is having some classes that operate on model objects and can be shared among display contexts. This is a new requirement due to the change of scope in display contexts: when they were created they were associated to a model object, but since the last commits, DCs now represent an UI screen.