LPS-63706 Publish portal-test* artifacts before building "core" modules
The "core" dir contains a few "*-test" modules, which need those artifacts in order to be built: the dependency is added by default in LiferayDefaultsPlugin and, if a module is a "*-test" one, building the jar requires building the test classes. We can bypass this problem in two ways: publishing the portal-test* artifacts, or excluding the "core/*-test" modules.