LPS-66136 Create util class to upgrade the plugin release
I needed to manually inject a reference to the counter service to avoid
a "BeanLocator not set" error. It seems not everything is fully
initilized when the upgrade registrators are invoked.

This class is responsible of initializing the `Release` table, creating
a row with version '0.0.1' if the portlet is in use. If no portlet is
found, no action is needed, as a '0.0.0' version will be implictly used
by the upgrade system.