LPS-66586 Handle WAB redeployment errors
Repositories deployed as WABs have some problems with hot deployment,
things aren't cleaned up correctly. Later on, if the user tries to do
anything with a repository, it will fail with a class cast exception.

While this doesn't fix the root cause, this change will at least allow
users to remount the repository and work around the problem.